moving database files to folder
