+++ databases/newsSqlDatabase.js
@@ -1,5 +1,5 @@
 var { Pool } = require('pg');
-var app = require('./app');
+var app = require('../app');
 
 var testString = 'postgres://jdkfzrdp:Rj2bM4jUptX89UEQEUiC0pyb6SMPZN6L@raja.db.elephantsql.com:5432/jdkfzrdp';
 var devString = 'postgres://xsuzfsll:sSa6qY83tfbIeykZ4Np1mROIohahR9w0@salt.db.elephantsql.com:5432/xsuzfsll';
@@ -75,17 +75,3 @@ const client = {
 }
 
 module.exports = client;
-
-/*
-module.exports = {
-  query: (text, callback) => {
-    const start = Date.now();
-    return pool.query(text, (err, res) => {
-      const duration = Date.now() - start;
-      const rows = res ? res.rowCount : undefined;
-      console.log('executed query', {text, duration, rows});
-      callback(err, res)
-    })
-  }
-};
-*/
